Some refactoring to allow for light updates to be executed asynchronously
