+++ src/pocketmine/level/ChunkManager.php
@@ -69,6 +69,48 @@ interface ChunkManager{
 	public function setBlockDataAt(int $x, int $y, int $z, int $data);
 
 	/**
+	 * Returns the raw block light level
+	 *
+	 * @param int $x
+	 * @param int $y
+	 * @param int $z
+	 *
+	 * @return int
+	 */
+	public function getBlockLightAt(int $x, int $y, int $z) : int;
+
+	/**
+	 * Sets the raw block light level
+	 *
+	 * @param int $x
+	 * @param int $y
+	 * @param int $z
+	 * @param int $level
+	 */
+	public function setBlockLightAt(int $x, int $y, int $z, int $level);
+
+	/**
+	 * Returns the highest amount of sky light can reach the specified coordinates.
+	 *
+	 * @param int $x
+	 * @param int $y
+	 * @param int $z
+	 *
+	 * @return int
+	 */
+	public function getBlockSkyLightAt(int $x, int $y, int $z) : int;
+
+	/**
+	 * Sets the raw block sky light level.
+	 *
+	 * @param int $x
+	 * @param int $y
+	 * @param int $z
+	 * @param int $level
+	 */
+	public function setBlockSkyLightAt(int $x, int $y, int $z, int $level);
+
+	/**
 	 * @param int $chunkX
 	 * @param int $chunkZ
 	 *
